Move song data types to MusicTypes file
